Added Earn page
